Add resources for Products and Categories

+++ lavastore_backend/app/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use App\Models\Product;
 use Illuminate\Http\Request;
 use App\Http\Requests\ProductCreateRequest;
 use App\Http\Requests\ProductUpdateRequest;
+use App\Http\Resources\ProductResource;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Http\Repositories\ProductRepository;
@@ -14,11 +15,8 @@ use Spatie\QueryBuilder\QueryBuilder;
 class ProductController extends Controller
 {
     use AuthorizesRequests;
-    
-    public function __construct(protected ProductRepository $productRepository)
-    {
-        $this->authorizeResource(Product::class, 'product');
-    }
+
+    public function __construct(protected ProductRepository $productRepository) {}
 
     public function index(Request $request): JsonResponse
     {
@@ -31,17 +29,18 @@ class ProductController extends Controller
 
         return response()->json([
             'message' => 'Products retrieved successfully',
-            'data' => $products
-        ], 200);
+            'data' => ProductResource::collection($products)
+        ]);
     }
 
     public function store(ProductCreateRequest $request): JsonResponse
     {
+        $this->authorize('create', Product::class);
         $product = $this->productRepository->create($request->validated());
 
         return response()->json([
             'message' => 'Product created successfully',
-            'data' => $product
+            'data' => new ProductResource($product)
         ], 201);
     }
 
@@ -49,7 +48,7 @@ class ProductController extends Controller
     {
         $product = $this->productRepository->getById($id);
 
-        if($product == null) {
+        if (!$product) {
             return response()->json([
                 'message' => 'Product not found'
             ], 404);
@@ -57,22 +56,37 @@ class ProductController extends Controller
 
         return response()->json([
             'message' => 'Product retrieved successfully',
-            'data' => $product
-        ], 200);
+            'data' => new ProductResource($product)
+        ]);
     }
 
     public function update(ProductUpdateRequest $request, string $id): JsonResponse
     {
+        $product = $this->productRepository->getById($id);
+        if (!$product) {
+            return response()->json([
+                'message' => 'Product not found'
+            ], 404);
+        }
+        $this->authorize('update', $product);
         $product = $this->productRepository->update($id, $request->validated());
 
         return response()->json([
             'message' => 'Product updated successfully',
-            'data' => $product
+            'data' => new ProductResource($product)
         ]);
     }
 
     public function destroy(string $id): JsonResponse
     {
+        $product = $this->productRepository->getById($id);
+        if (!$product) {
+            return response()->json([
+                'message' => 'Product not found'
+            ], 404);
+        }
+        
+        $this->authorize('delete', $product);
         $this->productRepository->delete($id);
 
         return response()->json([
